+++ core/widgets/init.php
@@ -64,16 +64,8 @@ class Layers_Widgets {
 	public function register_sidebars(){
 		global $wp_customize , $temp_sidebars;
 
-		// Fetch Builder Pages
-		$pages = get_pages(array(
-			'post_status' => 'publish,draft,private',
-			'meta_key' => '_wp_page_template',
-			'meta_value' => 'builder.php',
-			'parent' => -1
-		));
-
 		// Loop the Builder Pages and create their sidebars
-		foreach($pages as $page){
+		foreach( layers_get_builder_pages() as $page){
 			$this->register_builder_sidebar( $page->ID, $page->post_title );
 		}
 	}
@@ -195,7 +187,7 @@ class Layers_Widgets {
 			true
 		);
 		wp_localize_script( LAYERS_THEME_SLUG . '-admin-slider-widget' , 'sliderwidgeti8n', array(
-        	'confirm_message' => __( 'Are you sure you want to remove this slide?' , 'layerswp' )
+			'confirm_message' => __( 'Are you sure you want to remove this slide?' , 'layerswp' )
 		) );
 
 		// Content Widget
@@ -207,7 +199,7 @@ class Layers_Widgets {
 			true
 		);
 		wp_localize_script( LAYERS_THEME_SLUG . '-admin-content-widget' , 'contentwidgeti8n', array(
-        	'confirm_message' => __( 'Are you sure you want to remove this column?' , 'layerswp' )
+			'confirm_message' => __( 'Are you sure you want to remove this column?' , 'layerswp' )
 		) );
 
 		// Tiny MCE Initiator