+++ header.php
@@ -3,7 +3,6 @@
 <head>
 	<meta charset="<?php bloginfo( 'charset' ); ?>">
 	<meta name="viewport" content="initial-scale=1,maximum-scale=1,user-scalable=no">
-	<title><?php wp_title( '|', true, 'right' ); ?></title>
 	<?php wp_head(); ?>
 </head>
 <body <?php body_class(); ?>>